+++ src/components/WeatherCard.js
@@ -1,42 +1,11 @@
 import React from "react";
+import { dateBuilder } from "../commons/utils";
 
 import "../styles/weather-card.scss";
 
 function WeatherCard({ data }) {
   // console.log(data);
 
-  const dateBuilder = (d) => {
-    const months = [
-      "January",
-      "February",
-      "March",
-      "April",
-      "May",
-      "June",
-      "July",
-      "August",
-      "September",
-      "October",
-      "November",
-      "December",
-    ];
-    const weekDays = [
-      "Sunday",
-      "Monday",
-      "Tuesday",
-      "Wednesday",
-      "Thursday",
-      "Friday",
-      "Saturday",
-    ];
-
-    let day = weekDays[d.getDay()];
-    let date = d.getDate();
-    let month = months[d.getMonth()];
-    let year = d.getFullYear();
-
-    return `${day} ${date} ${month} ${year}`;
-  };
   return (
     <div className="weather-card">
       <div className="location-box">
