+++ src/styles/home.scss
@@ -40,15 +40,6 @@
       transform: translate(-50%);
     }
 
-    .loading {
-      position: absolute;
-      display: block;
-      top: 85%;
-      left: 50%;
-      margin-left: -1.25em;
-      margin-top: -1.5em;
-    }
-
     .error-text {
       position: absolute;
       display: block;
